+++ src/gui.js
@@ -86,7 +86,7 @@ BlockVisibilityDialogMorph, ThreadManager*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.gui = '2021-November-12';
+modules.gui = '2021-November-14';
 
 // Declarations
 
@@ -4235,12 +4235,15 @@ IDE_Morph.prototype.settingsMenu = function () {
 
     addPreference(
         'HSL pen color model',
-        () => SpriteMorph.prototype.penColorModel =
-            SpriteMorph.prototype.penColorModel === 'hsl' ? 'hsv' : 'hsl',
+        () => {
+            SpriteMorph.prototype.penColorModel =
+                SpriteMorph.prototype.penColorModel === 'hsl' ? 'hsv' : 'hsl';
+            this.refreshIDE();
+        },
         SpriteMorph.prototype.penColorModel === 'hsl',
         'uncheck to switch pen colors\nand graphic effects\nto HSV',
-        'check to switch pen colors\nand graphic effects\nfrom to HSL',
-        true
+        'check to switch pen colors\nand graphic effects\n to HSL',
+        false
     );
 
     addPreference(
@@ -4731,7 +4734,7 @@ IDE_Morph.prototype.aboutSnap = function () {
         module, btn1, btn2, btn3, btn4, licenseBtn, translatorsBtn,
         world = this.world();
 
-    aboutTxt = 'Snap! 7 - dev211112 -\nBuild Your Own Blocks\n\n'
+    aboutTxt = 'Snap! 7 - dev211114 -\nBuild Your Own Blocks\n\n'
         + 'Copyright \u24B8 2008-2021 Jens M\u00F6nig and '
         + 'Brian Harvey\n'
         + 'jens@moenig.org, bh@cs.berkeley.edu\n\n'